+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/Combinator/SingleEra.hs
@@ -3,4 +3,3 @@ module Ouroboros.Consensus.HardFork.Combinator.SingleEra (
   ) where
 
 import           Ouroboros.Consensus.HardFork.Combinator.SingleEra.Combined as X
-import           Ouroboros.Consensus.HardFork.Combinator.SingleEra.Info as X
